move components into dedicated files
